requirements.txt
google-adk
google-genai
pillow
